# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Northstar Commerce Gateway
  version: 1.0.0
  description: Stable public contract composed from independently deployable commerce capabilities.
servers:
  - url: http://localhost:4000/api/v1
paths:
...
  /analytics/summary:
    get:
      operationId: getCommerceAnalytics
      description: Returns an eventually consistent aggregate produced by the Python RabbitMQ consumer.
      responses:
        "200":
          description: Commerce summary
          content: { application/json: { schema: { $ref: "#/components/schemas/AnalyticsSummary" } } }
        "502": { $ref: "#/components/responses/ServiceError" }
components:
  responses:
    ServiceError:
      description: Standard error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
...
    AnalyticsSummary:
      type: object
      required: [orders, revenue, averageOrderValue, items, generatedAt]
      properties:
        orders: { type: integer, minimum: 0 }
        revenue: { type: number, minimum: 0 }
        averageOrderValue: { type: number, minimum: 0 }
        items: { type: integer, minimum: 0 }
        generatedAt: { type: string, format: date-time }
    ErrorEnvelope:
      type: object
      required: [error, requestId]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object }
        requestId: { type: string }